Evolution-Todo/phase-3-ai-chatbot/backend/requirements.txt
fastapi==0.115.0
sqlmodel==0.0.22
uvicorn[standard]==0.32.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic>=2.10,<3
pydantic-settings==2.6.0
passlib[bcrypt]==1.7.4
cachetools>=5.5.0
bcrypt>=4.0.0 
psycopg2-binary>=2.9.11
email-validator==2.3.0
slowapi>=0.1.9
reportlab>=4.0.0
alembic>=1.13.0
python-json-logger>=2.0.7
cachetools>=5.5.0
asyncpg>=0.31.0
openai-agents<0.2.9
google-generativeai==0.8.4